only use containers, remove hoomd import from top of file
